Moved the listview to fragment Home
