+++ app/src/main/java/com/example/trading_simulator/MainActivity.java
@@ -33,38 +33,6 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-
-
-        swipeRefreshLayout = findViewById(R.id.simpleSwipeRefreshLayout);
-
-        final StockPriceAccesser accesser = new RandomStockPriceGenerator();
-        simpleListView=(ListView)findViewById(R.id.simpleListView);
-
-
-        showListView(accesser);
-
-
-        swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                // cancel the Visual indication of a refresh
-
-                showListView(accesser);
-                swipeRefreshLayout.setRefreshing(false);
-            }
-        });
-        //perform listView item click event
-        simpleListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                Toast.makeText(getApplicationContext(),stockName[i], Toast.LENGTH_LONG).show();//show the selected image in toast according to position
-            }
-        });
-
-
-
-
-
         BottomNavigationView navView = findViewById(R.id.nav_view);
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
@@ -74,27 +42,6 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
         NavigationUI.setupWithNavController(navView, navController);
-    }
-
-    /**
-     * Display the ListView of stocks and prices when refreshes
-     * @param accesser The accesser to the stock price
-     */
-    private void showListView(StockPriceAccesser accesser) {
-        ArrayList<HashMap<String,String>> arrayList=new ArrayList<>();
-        for (int i=0;i<stockName.length;i++)
-        {
-            HashMap<String,String> hashMap=new HashMap<>();//create a hashmap to store the data in key value pair
-            hashMap.put("name",stockName[i]);
-            hashMap.put("price",String.format("$%.2f", accesser.getCurrentPrice(stockName[i])));
-            arrayList.add(hashMap);//add the hashmap into arrayList
-        }
-        System.out.println("Refreshed");
-        String[] from={"name","price"};//string array
-        int[] to={R.id.name,R.id.price};//int array of views id's
-        SimpleAdapter simpleAdapter=new SimpleAdapter(this,arrayList,R.layout.list_view_items,from,to);//Create object and set the parameters for simpleAdapter
-        simpleListView.setAdapter(simpleAdapter);//sets the adapter for listView
-
     }
 
 }
